+++ src/utils/extramath.h
@@ -5,12 +5,12 @@
 #define min(a, b) ({ (a < b) ? a : b; })
 #define max(a, b) ({ (a > b) ? a : b; })
 
-inline double clamp(double a, double min_, double max_) {
-	double m = max(a, min_);
+inline float clamp(float a, float min_, float max_) {
+	float m = max(a, min_);
 	return min(m, max_);
 }
 
-inline double smoothstep(double a, double b, double u) {
-	double t = clamp((u - a) / (b - a), 0.0, 1.0);
+inline float smoothstep(float a, float b, float u) {
+	float t = clamp((u - a) / (b - a), 0.0, 1.0);
 	return t * t * (3 - 2 * t);
 }
